Move to helper function and test
